requirements.txt

google-adk
google-genai

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

asyncio
python-dotenv
